Disallow browser-based signed fetch
The rationale for disallowing the Signature header is that if a browser-based
client can sign a fetch request, the end-user has the signing key, and can
issue `Create` or other activities without them being screened by the domain
holding their actor.

Its fine for single-user instances, but our security model assumes an actor
can create any object identifier within the same domain, and therefore that
the remote instance is responsible for not allowing its actors to spoof other
actor's object ids. Therefore we cannot safely support multi-user instances
which give the signing keys to the end-users.

This has the unwelcome effect of making the CORS header unsufficient for
private posts and instances with authorized fetch mode, though.