Change index_accounts_on_uri to text_pattern_opts to improve queries
`accounts` are searched for exact `uri` match as well as prefix matches
(in `remote_followers_hash`). Switching to `text_pattern_opts` will not
change anything to the former, but will make the index usable for the
latter.

The speedup should only matter to accounts with large amounts of remote
followers, though.