Fix inline image rendering
For posts with inline images, we need to set a `max-height` so that high-resolution images don't take up too much vertical space when rendering. This only affects images in `.status__content__text`, so it doesn't affect image galleries at all. Also adding a margin to the bottom for nicer spacing.

Fixes #1145.