Use FeaturedTag serializer for featured_tags/suggestions
This should be safe since the only relevant fields in a Tag (as far as suggestions for featuring a tag) would be the `name` and maybe the `url`. However, the inclusion of `history` is misleading as this represents the history of *everyone* using the tag, not just you. Also, it is completely irrelevant whether you are `following` the tag or not, since we are only considering your own posts (and your own posts are included in the home feed).

More generally, it makes way more sense to be working with FeaturedTags when asking for suggestions on which tags to feature. The `url` is localized to your own account, so you can easily view tagged posts from yourself only.