adds OpenTelemetry to the mix
- Only attempts to load and configure OpenTelemetry
  if an OTel endpoint has been provided.
- Some shenanigans to set the service name according
  to the names used in Procfile and docker-compose.

- disable ActiveRecord instrumentation

AR instrumentation currently prepends ActiveRecord#reload which pairs
catastrophically with attr_encrypted's method alias/rename and
monkeypatch. Together they call each other in a loop resulting in a
stack overflow.

Emit AR spans with an ActiveSupport instrumentation subscriber. Hand
the subscriber a new Tracer, specific to this subscription so that it
reports the "scope" of the ActiveRecord AS::Notifications as coming
from ActiveRecord.