Separate streaming node_modules from asset compilation
Applying this change shrinks docker image from 1.72GB to 1.09GB (-37%).

The gist of the change is that previously we kept `node_modules` used
for static asset compilation intermixed with `node_modules` needed
for the streaming service. Separating those allows us to keep just
the latter, since the former is not needed at runtime.

I also split build steps in the build image. This doesn't change
the number of layers in the resulting image, but it allows for
better local caching for build iteration.