Force links to be rendered without taking any RTLO into account
Avoids an attack where profile URLs can be constructed that trick
the user into thinking they have a different target. This forces the
browser to render the link itself, ignoring any nearby RTLO characters.
This cannot be used to fake verification (as the whole field has to be
only a URL, so the RTLO character stops verification), but it can be
used to have confusing profile links.