Allow streaming to connect to postgress with self-signed certs
Currently, if you run your mastadon instance on e.g. DigitalOcean and
use their managed postgres offering for the database, in order to be
able to connect you need to manually patch streaming/index.js in order
for it to be able to talk to the database.

Here is an example patch: https://github.com/rasca/mastodon/commit/f6d6bf19b59bb2d27ac7d3499466fc64c6503ef9

Now it turns out that the problem is that we're rolling our own
postgres connection string parsing, instead of using the one provided by
the library, which actually suppors everything we do and more and also
happens to be signature-compatible with our own function, so it seems
trivial to replace.

After this patch, it should be possible to simply properly set the
configuration string to e.g.
```
DATABASE_URL=postgresql://user:password@host:port/mastodon_production?sslmode=no-verify&ssl=true
```
and drop the (fragile) workarounds.

That said, it should be noted, that DigitalOcean provides the conenction
string with `sslmode=require`, so it should still be documented
somewhere that the users of this need to change it to be
`sslmode=no-verify` instead