Add a setting to remind user to add alt text.
When a user posts a toot with an image, we want to remind them to add
alt text. They can disable this reminder if they want, but by default
it'll be enabled.

This commit implements the setting itself, but doesn't add the warning
modal on the frontend yet.

The idea will be that, when hitting "Post" in the Mastodon UI with a
status that has one or more images attached, the frontend will check
whether all images on the status has been given alt text. And if not,
it will render a warning modal allowing the user to either cancel
posting the status so they can add alt text, or post the status anyway.

See #19974.