Add warning modal when creating post with images.
When attempting to publish a post with a media file that is missing alt
text, you'll now get a warning (unless you've disabled this warning in
the settings).

Only problem now is that the "Publish anyway" button doesn't actually
work currently, and also I need to replace the placeholder text in the
modal. So I need to figure that part out.

See #19974.