Add `id` attribute to REST::TagSerializer.
This was missing the `id` attribute, and is used in the
`api/v1/admin/trends/tags` endpoint, where it's necessary for app devs
to impliment admin functionality.

The controller should maybe be updated to use the Admin-specific
TagSerializer, but that'd be a breaking change so I'm just doing this
for now.