Add exclusive lists
This adds a checkbox to the list editor modal that lets you say a list is "Exclusive". If this is checked, the posts that appear in this list timeline do not get rendered in the user's Home feed.

The utility is explained here: https://github.com/hometown-fork/hometown/wiki/Exclusive-lists

This adds an is_exclusive boolean column to the lists table, modifies the feed manager and various list editor states.

TODO: internationalization, UI review