app/workers: add JitteredRetryIn, use in UnfollowFollowWorker
This adds the JitteredRetryIn concern, which adds a random amount of
jitter to the retry time for a failed job. The implementation matches
that of the default Sidekiq retry time, plus an additional jitter of up
to 50% of the retry time.

We then use this in UnfollowFollowWorker to jitter retries; this should
help with cases where users with many followers migrate to a smaller
server. Without jitter, hundreds or thousands of jobs are scheduled at
the same time, some number of them succeed or fail, and then the failed
ones all retry at exactly the same time, DoS-ing the target server.