Disable ElasticSearch for AccountSearchService spec
Out of the box, if ElasticSearch is available, the AccountSearchService spec will try to use it, and some of the examples will fail. This change explicitly disables Chewy in a localized way.

I'm not sure this is the RIGHT way to do it. I don't think any of the specs rely on ES, so it might make more sense to globally disable Chewy in the spec helper?