Move Ruby code linting back to Codeclimate
Switching to super-linter reveals too many linting errors in files I
touch in PRs.

It seems that the PR that did this: https://github.com/mastodon/mastodon/pull/18587 -- is largely frontend-motivated, and that backend was better off with codeclimate.

Since both codeclimate and superlinter is configured in principle
anyway, we might as well have codeclimate for the backend and
superlinter for the frontend.