Allow configuring rate limits via environment variables (#15832)
This commit introduces the ability to set all existing `Rack::Attack` parameters using environment variables so that they can be configured by administrators before the service starts up.

I chose environment variables because:

* It is the least possible thing that could work.
* It is easy to configure *before* starting up the service, making it harder to circumvent the limits while Rails is still initializing (though I don't believe this is a real concern).
* It is easy to configure in Docker, etc.

I have created the ability to configure *all* rate limits, above and beyond those which are read-only, in order to use a consistent pattern in the code, but I recognize that this makes it very slightly easier for administrators to create rogue, misconfigured instances which can swamp others. If we find the risk is too great, we can remove those environment variables.

Personally, I think there are a few different implementations that might be better, but I wanted to float the idea in the form of a pull request to start a conversation about what kind of implementation would be most acceptable. For this reason, I have not yet written new tests, nor updated any documentation. If we align on this (or some other) implementation, I can add those in a new commit.

See https://github.com/mastodon/mastodon/issues/15832.