yarn.lock: add hash so NixOS fetching works (#1283)
NixOS uses the hashes in yarn.lock to fetch dependencies at build time;
this dependency doesn't have a hash, so building Hometown on NixOS would
fail. Adding a hash (calculated manually with curl + shasum) fixes the
build on NixOS and doesn't change the behaviour otherwise.

This broke in #17902 upstream previously; this is a similar fix here.