Fix counter obfuscation logic
This fixes an issue that obfuscates reply, boost, and favourite counts by default on the web app. Visually, it gives users the misleading impression that there is less engagement/less activity on the Mastodon network than is actually taking place. It also creates a discordant experience for users of 3rd-party clients such as Ivory, IceCubes, and Elk-- all of which display the accurate counts.