Fix paths with url-encoded @ to redirect to the correct path
Some web platforms insist to urlencode the '@' sign in URLs. Toots,
profile urls etc. pasted in such a platform wil generate hyperlinks to
Mastodon with the '@' encoded to '%40', yielding a 404 on Mastodon.

In order to not break these links for users, redirect any incoming
URLs where the @ is encoded to the correct unencoded version, for
further processing.

Closes: #15153