Fix single-record invalid condition on PollVote
Previously the logic in the VoteValidator would consider the presence of
any votes (or any votes with same choice on multiple) as sufficient to
mark a record invalid -- even if the only vote in question was a saved
vote that was the same record as the one being checked for validity.

This would lead to a scenario where a saved PollVote would consider
itself invalid even if it was the only vote by that account on that
poll.

This change updates the logic to have already persisted records check
for votes that are not the same as themselves, and preserves the
previous behavior for non-persisted records being saved for the first
time.

Also added some spec coverages for various validity states of a vote,
and extracted some well named methods in the VoteValidator class.