Remove rewhere from AccountFilter and check params instead
In a scenario where we have both `origin=remote` and a value for
`by_domain` we can just ignore the `origin` param because its condition
will be taken care of by the domain check.

In scenarios where either origin is local or origin is remote but there
is not a domain value provided, params will be left in place as before.