Remove references to non-existent actions
Removes before_action references to two `show` actions which don't
exist, and to a route which points to a non-existent action.

The admin/domain_blocks#show action was removed in bd53dd521064b12261b82105624cf5f8b9ca9d69
but the route was not removed and the before_action wasn't changed.

The admin/email_domain_blocks#show and #destroy have never existed, and
may have been erroneously copy/pasted in when created.