Move link header setting to after_action
Because of an interaction with `preload_pack_asset` (called from
web_app partial) and the setting of these link headers, there's an
error in the rails code sending the link headers:

https://github.com/rails/rails/blob/v7.0.4.3/actionview/lib/action_view/helpers/asset_tag_helper.rb#L558

When these headers are set in a before action which is before the views
are rendered, they are present in the response['Link'] when that code
runs, and it errors out while trying to send the link headers related
to the preload of assets.

Moving to the before action means that they are still in the http
response sent back, but they aren't there yet when the view renders,
and so this code runs without error.