Reset model in migration after removing column
Migrations were failing to run all the way through (start with empty db,
run db:migrate) because the `Account` class still had the `silenced`
column as part of the schema when the `AddInstanceActor` migration tried
to run the `Account.create` line in the up part of the migration.

Previously the classes must have been reloaded between migrations, but
now they are staying cached between migrations.

This change explicitly resets the column information after the removal
of column in the earlier migration.

It looks like this is the only migration which:

- Has a class definition inside the migration
- Removes a column in the `up` method
- That class is used in later migrations to insert data

So this is likely the only correction to make in a previous migration.