Add a MigrationWarning helper for long migrations
Several migrations have a "Warning" in place to allow admins to pause or
cancel migrations which might be long-running.

This change extracts the announcement of that warning message to a new
module, and includes it in the migrations that make such announcements.

The text output and behavior of the warning should be unchanged.