Better explanation in onboarding text
'Some profiles' is incorrect, that really depends on the server size of the user. I don't have statistics about how many remote accounts users are following by average. In the 6 years I'm here most of my followers are on a remote server, but because we don't know for all people I made this text agnostic.