fix: remove port 80 mapping in devcontainer
This change leaves ports 3000 and 4000 mapped, but removes port 80.

On MacOS and possible other OSes, mapping port 80 requires a preference
change on docker as it's a privileged port. The devcontainer fails to
run without this change, which is a annoying. Additionally, this mapping
will fail for anyone that already has port 80 mapped for some other
purpose.