Add templating to mastodon-streaming systemd, maintaining backwards compatibility
The changes here introduce the templated version of mastodon-streaming, but retains a non-templated version which invokes the templated version with the default port. This means that you can do `sudo systemctl start mastodon-streaming` and get a single instance on port 4000, but also that you can separately start many instances on different ports with `sudo systemctl start mastodon-streaming@{4000,4001,4002}` if desired. Those instances can be stopped and restarted with `sudo systemctl restart mastodon-streaming` but cannot be started similarly.