Fix UI crash in moderation interface when opening the media modal
A dependency on redux was introduced to the media modal in #23891, but
this media modal is used in places were the redux store is not set up,
causing the JS to crash.

This commit partially reverts #23891 to remove the media modal's dependency
on redux, at the cost of `lang` attributes in that modal.